# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ruko Backend",
    "description": "Multilingual investor-protection API for suspicious investment tips",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/check/image": {
      "post": {
        "tags": [
          "Media Ingest"
        ],
        "summary": "Check Image",
        "description": "Ingest an image/screenshot, extract text via vision OCR in-memory, and return full CheckResult.",
        "operationId": "check_image_v1_check_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_check_image_v1_check_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/check/audio": {
      "post": {
        "tags": [
          "Media Ingest"
        ],
        "summary": "Check Audio",
        "description": "Ingest an audio voice note, transcribe text via Sarvam STT in-memory, and return full CheckResult.",
        "operationId": "check_audio_v1_check_audio_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_check_audio_v1_check_audio_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/check/media": {
      "post": {
        "tags": [
          "Media Ingest"
        ],
        "summary": "Check Media",
        "description": "Unified multimodal ingestion endpoint: sniffs magic bytes and routes to text, OCR, STT, or video pipeline.",
        "operationId": "check_media_v1_check_media_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_check_media_v1_check_media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_check_audio_v1_check_audio_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Audio voice-note file"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Optional language hint (hi, gu, en)"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount",
            "description": "Optional investment amount mentioned"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_check_audio_v1_check_audio_post"
      },
      "Body_check_image_v1_check_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Screenshot or photo of suspicious message"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Optional language hint (gu, hi, en)"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount",
            "description": "Optional investment amount mentioned"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_check_image_v1_check_image_post"
      },
      "Body_check_media_v1_check_media_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Multipart file: text (.txt), image, audio, or video reel"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Optional language hint (gu, hi, en)"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount",
            "description": "Optional investment amount mentioned"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_check_media_v1_check_media_post"
      },
...
      "CheckResult": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Unique request UUID"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Detected or requested language (gu, hi, en, unknown)"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict",
            "description": "strong_red_flags | cannot_verify | no_red_flags_found | out_of_scope"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Combined composite score from 0.0 to 1.0"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "type": "array",
            "title": "Reasons",
            "description": "Ordered reasons for verdict"
          },
          "registry": {
            "$ref": "#/components/schemas/RegistryInfo",
            "description": "Registry snapshot lookup information"
          },
          "model": {
            "$ref": "#/components/schemas/ModelInfo",
            "description": "ML model output details"
          },
          "claims": {
            "type": "object",
            "title": "Claims",
            "description": "Extracted claims"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "description": "Localized honest summary note"
          },
          "degraded": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degraded",
            "description": "List of degraded or unavailable modules"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "description": "Localized legal/educational disclaimer"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Extracted message text (for media OCR/STT/video)"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Input source (ocr | stt | text | video)"
          },
          "input_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Source",
            "description": "Normalized input source: text | ocr | stt | video",
            "default": "text"
          },
          "speech_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speech Text",
            "description": "Transcribed audio speech from video"
          },
          "on_screen_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Screen Text",
            "description": "Extracted on-screen text from video frames"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint",
            "description": "Contextual user guidance hint, e.g. for Instagram links"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "language",
          "verdict",
          "score",
          "registry",
          "model",
          "note",
          "disclaimer"
        ],
        "title": "CheckResult",
        "description": "Unified verdict and explainable report returned by Ruko."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelInfo": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available",
            "description": "Whether model inference was available"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "Model scam risk probability (0.0 to 1.0)"
          },
          "top_words": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Words",
            "description": "Words that increased scam risk score"
          },
          "calming_words": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Calming Words",
            "description": "Words that decreased scam risk score"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "ModelInfo",
        "description": "Machine learning model prediction details."
      },
...
      "Reason": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Unique machine-readable reason code"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "high | medium | low | info"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Localized explanation sentence"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence",
            "description": "Exact quoted phrase from the message"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "rule | model | registry"
          }
        },
        "type": "object",
        "required": [
          "code",
          "severity",
          "text",
          "evidence",
          "source"
        ],
        "title": "Reason",
        "description": "Detailed reason for a red flag or finding."
      },
      "RegistryInfo": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "found_in_snapshot | possible_match | not_found_in_snapshot | not_checked"
          },
          "matches": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Matches",
            "description": "Matching records from snapshot"
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date",
            "description": "Date of local registry snapshot",
            "default": "2026-03-01"
          },
          "is_sample_data": {
            "type": "boolean",
            "title": "Is Sample Data",
            "description": "True if sample/demo data is used",
            "default": true
          },
          "verify_url": {
            "type": "string",
            "title": "Verify Url",
            "description": "Official verification URL"
          }
        },
        "type": "object",
        "required": [
          "status",
          "verify_url"
        ],
        "title": "RegistryInfo",
        "description": "Snapshot verification outcome for claimed registration numbers/names."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}